feat: add private symbols scaffolding to PhelFnNormalizer
